+++ authelio_sdk_tests/integration/tests/group/test_filter.py
@@ -1,30 +1,26 @@
 from typing import Callable
 
-from faker import Faker
-
 from authelio_sdk.client import Client
 from authelio_sdk.models.group import Group
 
 
 def test_FUNC_client_group_filter_WITH_new_group_EXPECT_new_group_found(
         sdk_client: Client,
-        group_function: Callable[..., Group],
-        faker: Faker
+        group_function: Callable[..., Group]
 ) -> None:
     """
     Test whether filter function works as expected.
 
     :param sdk_client: Authelio SDK client.
     :param group_function: Function, that creates group entity.
-    :param faker: Faker fixture.
 
     :return: No return.
     """
     # Create some groups for filtering.
     group_ids = [group_function().group_id for _ in range(5)]
 
-    filtered_groups = sdk_client.group.filter()
+    filtered_groups, next_page_id = sdk_client.group.filter()
 
-    filtered_group_ids = [key for key in filtered_groups.keys()]
+    filtered_group_ids = list(filtered_groups.keys())
 
-    assert filtered_group_ids
+    assert all(group_id in filtered_group_ids for group_id in group_ids)